#34: Rename wrongly named packages * renamed package `naming_conventions` to `naming` and reorganized the files within * renamed package `statushandling` to `status` and reorganized the files within * renamed `DBSeqFunction` to `DBMultipleResultFunction` * renamed `DBUniqueFunction` to `DBSingleResultFunction` * renamed `DBOptionFunction` to `DBOptionalResultFunction` * renamed the execution methods within `DBFunction` to match the above successor names
