+++ core/src/main/scala/za/co/absa/fadb/DBEngine.scala
@@ -45,7 +45,7 @@ trait DBEngine {
     * @tparam R     - return the of the query
     * @return       - sequence of the results of database query
     */
-  def execute[R](query: QueryType[R]): Future[Seq[R]] = run(query)
+  def fetchAll[R](query: QueryType[R]): Future[Seq[R]] = run(query)
 
   /**
     * Public method to execute when query is expected to return exactly one row
@@ -53,7 +53,7 @@ trait DBEngine {
     * @tparam R     - return the of the query
     * @return       - sequence of the results of database query
     */
-  def unique[R](query: QueryType[R]): Future[R] = {
+  def fetchHead[R](query: QueryType[R]): Future[R] = {
     run(query).map(_.head)
   }
 
@@ -64,7 +64,7 @@ trait DBEngine {
     * @return       - sequence of the results of database query
     */
 
-  def option[R](query:  QueryType[R]): Future[Option[R]] = {
+  def fetchHeadOption[R](query:  QueryType[R]): Future[Option[R]] = {
     run(query).map(_.headOption)
   }
 }
