+++ core/src/main/scala/za/co/absa/fadb/DBSchema.scala
@@ -16,7 +16,7 @@
 
 package za.co.absa.fadb
 
-import za.co.absa.fadb.naming_conventions.NamingConvention
+import za.co.absa.fadb.naming.NamingConvention
 
 /**
   * An abstract class, an ancestor to represent a database schema (each database function should be placed in a schema)
@@ -25,7 +25,7 @@ import za.co.absa.fadb.naming_conventions.NamingConvention
   *                           possibility of override
   * @param dBEngine           - [[DBEngine]] to execute the functions with. Not directly needed for the DBSchema class, rather
   *                           to be passed on to [[DBFunction]] members of the schema
-  * @param namingConvention   - the [[za.co.absa.fadb.naming_conventions.NamingConvention NamingConvention]] prescribing how to convert a class name into a db object name
+  * @param namingConvention   - the [[za.co.absa.fadb.naming.NamingConvention NamingConvention]] prescribing how to convert a class name into a db object name
   */
 abstract class DBSchema(schemaNameOverride: Option[String] = None)
                        (implicit dBEngine: DBEngine, implicit val namingConvention: NamingConvention) {
@@ -57,7 +57,7 @@ abstract class DBSchema(schemaNameOverride: Option[String] = None)
 
   /**
     * Function to convert a class to the associated DB object name, based on the class' name. For transformation from the
-    * class name to usual db name the schema's [[za.co.absa.fadb.naming_conventions.NamingConvention NamingConvention]] is used.
+    * class name to usual db name the schema's [[za.co.absa.fadb.naming.NamingConvention NamingConvention]] is used.
     * @param c  - class which name to use to get the DB object name
     * @return   - the db object name
     */
