+++ core/src/test/scala/za/co/absa/fadb/DBSchemaSuite.scala
@@ -16,7 +16,7 @@
 package za.co.absa.fadb
 
 import org.scalatest.funsuite.AnyFunSuite
-import za.co.absa.fadb.naming_conventions.SnakeCaseNaming.Implicits.namingConvention
+import za.co.absa.fadb.naming.implementations.SnakeCaseNaming.Implicits.namingConvention
 
 import scala.concurrent.{Await, Future}
