+++ core/src/test/scala/za/co/absa/fadb/naming_conventions/SnakeCaseNamingSuite.scala
@@ -2,7 +2,8 @@ package za.co.absa.fadb.naming_conventions
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
-import za.co.absa.fadb.naming_conventions.letters_case.LettersCase._
+import za.co.absa.fadb.naming.implementations.SnakeCaseNaming
+import za.co.absa.fadb.naming.LettersCase._
 
 class SnakeCaseNamingSuite extends AnyWordSpec with Matchers {
   private class ThisIsATestClass
