+++ core/src/test/scala/za/co/absa/fadb/status/StatusExceptionSuite.scala
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package za.co.absa.fadb.statushandling
+package za.co.absa.fadb.status
 
 import org.scalatest.funsuite.AnyFunSuite
-import za.co.absa.fadb.statushandling.StatusException._
+import za.co.absa.fadb.status.StatusException._
 
 class StatusExceptionSuite extends AnyFunSuite {
   test("Test equals - when they are the same") {
