+++ core/src/test/scala/za/co/absa/fadb/status/StatusHandlingSuite.scala
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package za.co.absa.fadb.statushandling
+package za.co.absa.fadb.status
 
 import org.scalatest.funsuite.AnyFunSuite
 import za.co.absa.fadb.DBFunctionFabric
-import za.co.absa.fadb.naming_conventions.{NamingConvention, SnakeCaseNaming}
+import za.co.absa.fadb.naming.NamingConvention
+import za.co.absa.fadb.naming.implementations.SnakeCaseNaming
+import za.co.absa.fadb.status.handling.StatusHandling
 
 class StatusHandlingSuite extends AnyFunSuite {
   test("Fields to select filled with default values") {
