+++ examples/src/test/scala/za/co/absa/fadb/examples/enceladus/DatasetSchemaSuite.scala
@@ -21,7 +21,7 @@ import org.scalatest.wordspec.AnyWordSpec
 import za.co.absa.fadb.examples.enceladus.DatasetSchema._
 import slick.jdbc.PostgresProfile.api._
 import za.co.absa.fadb.slick.SlickPgEngine
-import za.co.absa.fadb.statushandling.StatusException
+import za.co.absa.fadb.status.StatusException
 
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
